tolerate base_distro check & WSL detect
This will not work on Debian Linux. The changed script will search for all values with ID in the key. If there is the required value in the returned string, it passes the test.
WSL detection is not working on my debian WSL. I found the answer here: https://stackoverflow.com/a/38859331/13406850. It has been applied to the function.